fix: userId not generated in the first visit
