+++ src/routes/+layout.server.ts
@@ -1,20 +1,5 @@
-import { generateUserId } from '$lib/generateUserId.js';
-
 export function load({ cookies }) {
 	const cactoideUserId = cookies.get('cactoideUserId');
-	const userId = generateUserId();
-
-	const DAYS = 400; // practical upper bound in many browsers for cookies
-	const MAX_AGE = 60 * 60 * 24 * DAYS;
-	const PATH = '/';
-
-	if (!cactoideUserId) {
-		console.debug(`There is no cactoideUserId cookie, generating new one...`);
-		cookies.set('cactoideUserId', userId, { path: PATH, maxAge: MAX_AGE });
-	} else {
-		console.debug(`cactoideUserId: ${cactoideUserId}`);
-		console.debug(`cactoideUserId cookie found, using existing one...`);
-	}
 
 	return {
 		cactoideUserId
